refactor class desing
